Clustering using Consul needs atleast & atmost 1 Bootstrap node
If using the "bootstrap" variable, Consul accepts atmost 1 Bootstrap node.
if using "bootstrap-expect", the user can specify a number of Nodes that
can be in Bootstrap mode.

This commit provides a simple flag to be used if the node is dedicated
as bootstrap node. This is a temporary change until we find a more
dynamic bootstrapping process.

Signed-off-by: Madhu Venugopal <madhu@socketplane.io>