Interface identification logic is changed from a timeout to indefinite
When the user lets the daemon to choose the best interface for the job
and if the daemon is not able to find one, it will try indefinitely till
it will fine one. This is important because for a multi-host scenario,
unless the daemon indentifies an interface to bind to, it will not be
albe to cluster with its peers.

If the user knows exactly which interface to bind with, they can use the
--iface=xxx option to bind to a specific interface.

Signed-off-by: Madhu Venugopal <madhu@socketplane.io>