Fixes #47 with the help of ecc caching support
The idea is to use ecc APIs as is. But the ECC will provide a caching
front-end when Consul daemon couldnt identify an appropriate interface
to bind to.
In order to provide single-node setup, the bootstrap node will not wait
indefinitely for a port to bind with. Rather, it will make use of the
offline cache in cases where consul/bonjour couldnt identify a port to
bind to. A Non-bootstrap node is still made to wait indefinitely for the
eligible port exactly the same way it works now.

Signed-off-by: Madhu Venugopal <madhu@socketplane.io>