Enabled network_test
1. network Create and Destroy apis calls a few netlink APIs that require
the sudo permission to run these tests.
If we cannot run these tests in sudo, it will incorrectly deem the tests
as failed. Hence, changes are made to the testing code to make sure we
dont fail for permission failure cases.
If our CI can be executed with sudo privilege, then the code-coverage
will automatically work better. But still the error condition must be
checked for the netlink APIs.

2. Removed consul dependencies by not initing the ecc datastore. It will
automatically pick the local cache and hence UT is self contained
(almost)

Signed-off-by: Madhu Venugopal <madhu@socketplane.io>