Docker Compose with socketplane, docker-ovs and powerstrip containers
Please note that docker-compose cfgs are not a very programmable entity.
In order to use it effectively (such as `modprobe openvswitch`), we need
to have some external logic in the script that compensates for the lack
of programmability in docker-compose config files.

This commit doesn't address the scripting changes. It will be addressed
in subsequent commits.

Signed-off-by: Madhu Venugopal <madhu@socketplane.io>